CoupledIsingModels

Documentation for CoupledIsingModels.

CoupledIsingModels.AbstractBoundaryKind

CoupledIsingModels.AbstractComputation

CoupledIsingModels.AbstractEnergy

CoupledIsingModels.AbstractField

CoupledIsingModels.AbstractIsingModel

CoupledIsingModels.AbstractLatticeIsingModel

CoupledIsingModels.AbstractMagnetization

CoupledIsingModels.AbstractMeanField

CoupledIsingModels.AbstractMeanIsingAlgorithm

CoupledIsingModels.AbstractParallelUpdate

CoupledIsingModels.AbstractPeriodicBoundary

CoupledIsingModels.AbstractSerialUpdate

CoupledIsingModels.AbstractSpinGlassIsingModel

CoupledIsingModels.AbstractTemporalAverageAlgorithm

CoupledIsingModels.AbstractUpdateIsingModel

CoupledIsingModels.CheckerboardGlauberAlgorithm

CoupledIsingModels.CheckerboardMetropolisAlgorithm

CoupledIsingModels.InteractionEnergyField

CoupledIsingModels.LatticeIsingModel

CoupledIsingModels.LatticeIsingModel

CoupledIsingModels.LatticeIsingModel

CoupledIsingModels.LocalEnergyField

CoupledIsingModels.MagnetizationField

CoupledIsingModels.MeanFields

CoupledIsingModels.PeriodicBoundary

CoupledIsingModels.SpinGlassIsingModel

CoupledIsingModels.SpinGlassIsingModel

CoupledIsingModels.SpinGlassIsingModel

CoupledIsingModels.TemporalAverageAlgorithm

CoupledIsingModels.CheckerboardGlauberUpdate!

CoupledIsingModels.CheckerboardMetropolisUpdate!

CoupledIsingModels.GlauberUpdate!

CoupledIsingModels.GlauberUpdate!

CoupledIsingModels.LittleUpdate!

CoupledIsingModels.LittleUpdate!

CoupledIsingModels.MetropolisHastingsUpdate!

CoupledIsingModels.MetropolisHastingsUpdate!

CoupledIsingModels._TAPEquation

CoupledIsingModels._generatePartitionFunction

CoupledIsingModels._generateSpinCombinations

CoupledIsingModels._magnetization1DLatticeModel

CoupledIsingModels._magnetization2DLatticeModel

CoupledIsingModels.bin2dec

CoupledIsingModels.calcField

CoupledIsingModels.calcField

CoupledIsingModels.calculateFieldsAverage!

CoupledIsingModels.crossentropy

CoupledIsingModels.energy

CoupledIsingModels.energy

CoupledIsingModels.energy

CoupledIsingModels.energy

CoupledIsingModels.energy

CoupledIsingModels.energy

CoupledIsingModels.entropy

CoupledIsingModels.magnetization

CoupledIsingModels.magnetization

CoupledIsingModels.makeNearestNeighboursConnections

CoupledIsingModels.mutualInformation

CoupledIsingModels.probDist

CoupledIsingModels.sigmoid

CoupledIsingModels.spinProbabilities

CoupledIsingModels.AbstractBoundaryKind — Type

Abstract type for Kind of Boundary

CoupledIsingModels.AbstractComputation — Type

Abstract type for accelerating the computation

CoupledIsingModels.AbstractEnergy — Type

Abstract Energy

CoupledIsingModels.AbstractField — Type

Abstract Field

CoupledIsingModels.AbstractIsingModel — Type

The AbstractIsingModel has a 
T type of the arguments, in general T <: Number
N the dimension of the spin array, so N is an integer number
M enumerates the different spin values, so M is an integer number.
M = 2 means two values for the spin, -T(1) and T(1) for example.
M = 3 means three values for the spin, -T(1), T(0), T(1) for example.

CoupledIsingModels.AbstractLatticeIsingModel — Type

Abstract type for Lattice Ising models, the parameters are similar to
AbstractIsingModel{T, N, M}
in general we expect to have M = 2

CoupledIsingModels.AbstractMagnetization — Type

Abstract Magnetization

CoupledIsingModels.AbstractMeanField — Type

Abstract Mean Field

CoupledIsingModels.AbstractMeanIsingAlgorithm — Type

Abstract Mean for Ising Models

CoupledIsingModels.AbstractParallelUpdate — Type

Abstract type for parallel updating

CoupledIsingModels.AbstractPeriodicBoundary — Type

Abstract Periodic Boundary

CoupledIsingModels.AbstractSerialUpdate — Type

Abstract type for serial updating

CoupledIsingModels.AbstractSpinGlassIsingModel — Type

Abstract type for Spin Glass Ising models, the parameters are similar to
AbstractIsingModel{T, N, M}
in general we expect to have N = 1 and M = 2

CoupledIsingModels.AbstractTemporalAverageAlgorithm — Type

Abstract Temporal Average

CoupledIsingModels.AbstractUpdateIsingModel — Type

Abstract type for algorithms that update Ising  Models

CoupledIsingModels.CheckerboardGlauberAlgorithm — Method

CheckerboardGlauberAlgorithm
it is defined for Lattice Ising models
right now considers that in each dimension the ising model has an even size

CoupledIsingModels.CheckerboardMetropolisAlgorithm — Method

CheckerboardMetropolisAlgorithm
it is defined for Lattice Ising models
right now considers that in each dimension the ising model has an even size

CoupledIsingModels.InteractionEnergyField — Type

for the interaction energy field

CoupledIsingModels.LatticeIsingModel — Type

Lattice Ising Model with a simple spin chain
the Hamiltonian is given by 
H(s_1, s_0) = - ( H + J' * s_0)' * s_1 =  - ( H' * s_1 + s_0' * J * s_1 )
where in general s_1 is at time t=1 and s_0 is at time t=0

CoupledIsingModels.LatticeIsingModel — Method

# number type, in general a kind of Float 
T
# possible spin values
_s::SVector{M,T}
# shape of the spin array
shp::NTuple{N, Int}  
# random number generator
rng::AbstractRNG

J and H are filled with zeros of type T 
s is filled with random values drawn from _s

CoupledIsingModels.LatticeIsingModel — Method

# number type 
T
# shape of the spin array
shp::NTuple{N, Int}  

_s is assumed to be (T(-1), T(1))
J and H are filled with zeros of type T 
s is filled with random values drawn from _s

CoupledIsingModels.LocalEnergyField — Type

for the total local energy field

CoupledIsingModels.MagnetizationField — Type

For the magnetization field

CoupledIsingModels.MeanFields — Type

Mean Field

CoupledIsingModels.PeriodicBoundary — Type

Periodic Boundary Kind

CoupledIsingModels.SpinGlassIsingModel — Type

Spin Glass Ising Model with a simple spin chain
the Hamiltonian is given by 
H(s_1, s_0) = - ( H + J' * s_0)' * s_1 =  - ( H' * s_1 + s_0' * J * s_1 )
where in general s_1 is at time t=1 and s_0 is at time t=0

CoupledIsingModels.SpinGlassIsingModel — Method

# number type, in general a kind of Float 
T
# possible spin values
_s::SVector{M,T}
# size of the spin array
sze::int  
# random number generator
rng::AbstractRNG

J and H are filled with zeros of type T 
s is filled with random values drawn from _s

CoupledIsingModels.SpinGlassIsingModel — Method

# number type 
T
# size of the spin array
sze::int  

_s is assumed to be (T(-1), T(1))
J and H are filled with zeros of type T 
s is filled with random values drawn from _s

CoupledIsingModels.TemporalAverageAlgorithm — Type

For taking temporal Averages

CoupledIsingModels.CheckerboardGlauberUpdate! — Method

Glauber Checkerboard update

CoupledIsingModels.CheckerboardMetropolisUpdate! — Method

Metropolis Checkerboard update

CoupledIsingModels.GlauberUpdate! — Method

Glauber update

CoupledIsingModels.GlauberUpdate! — Method

Glauber update

CoupledIsingModels.LittleUpdate! — Method

Little parallel update
See:
Collective Properties of Neural Networks: A Statistical Physics Approach - P. Peretto
section 5

CoupledIsingModels.LittleUpdate! — Method

Little parallel update
See:
Collective Properties of Neural Networks: A Statistical Physics Approach - P. Peretto
section 5

CoupledIsingModels.MetropolisHastingsUpdate! — Method

Metropolis - Hastings update

CoupledIsingModels.MetropolisHastingsUpdate! — Method

Metropolis - Hastings update

CoupledIsingModels._TAPEquation — Method

TAP equation for magnetization
    m_i = tanh(β H_i + β J_ij*m_j - β² m_i (J_ij)^2*(1 - m_j^2)  ) )
see for example https://www.nature.com/articles/s41467-021-20890-5
equation (20)

CoupledIsingModels._generatePartitionFunction — Method

Generates the partition function and probabilities for a given ising model
consider that size needs to be at most 7

CoupledIsingModels._generateSpinCombinations — Method

Generates all spin combinations for a given chain 
of size sze and posible spins _spins

CoupledIsingModels._magnetization1DLatticeModel — Method

Magnetization in Lattice Ising Model in 1D with 
first neighbours interactions 
see for example https://www.thphys.uni-heidelberg.de/~wolschin/statsem20_3s.pdf
equation 3.27

CoupledIsingModels._magnetization2DLatticeModel — Method

Spontaneous Magnetization in Lattice Ising Model in 2D with first neighbours interactions see for example https://en.wikipedia.org/wiki/Isingmodel#Twodimensions equation in Onsager's formula for spontaneous magnetization

CoupledIsingModels.bin2dec — Method

Makes a base 10 number from binary representation

CoupledIsingModels.calcField — Method

energy

CoupledIsingModels.calcField — Method

magnetization

CoupledIsingModels.calculateFieldsAverage! — Method

calculate fields average

Inverse temperature
    Beta::T
Ising Model
    ising::AbstractIsingModel
Mean Fields to find
    meanFields:: MeanFields

CoupledIsingModels.crossentropy — Method

Calculates the cross entropy of two mass distributions

CoupledIsingModels.energy — Method

instant interaction local energy
for Serial Update methods

CoupledIsingModels.energy — Method

instant energy
for Parallel Update
TODO
see eq 38 of P. Peretto 1984 "Collective Properties of Neural Networks: A Statistical Physicis Approach"

CoupledIsingModels.energy — Method

instant total local energy
for Serial Update methods

CoupledIsingModels.energy — Method

instant interaction local energy
for Serial Update methods

CoupledIsingModels.energy — Method

instant energy
for Parallel Update
TODO
see eq 38 of P. Peretto 1984 "Collective Properties of Neural Networks: A Statistical Physicis Approach"

CoupledIsingModels.energy — Method

instant total local energy
for Serial Update methods

CoupledIsingModels.entropy — Method

Calculates the entropy of a mass distribution

CoupledIsingModels.magnetization — Method

instant magnetization

CoupledIsingModels.magnetization — Method

instant magnetization

CoupledIsingModels.makeNearestNeighboursConnections — Method

# distance to consider nearest neighbours, (NOTE is exact distance) 
d::Int
# boundary condition for each dimension N
boundaries::NTuple{N, AbstractBoundaryKind}
# shape of the spin array
shp::NTuple{N, Int}

CoupledIsingModels.mutualInformation — Method

Calculates the Mutual Information for a given set of spins, uses 
the results from probDist

CoupledIsingModels.probDist — Method

Calculates the probability distribution for a given set of spins, uses 
the results from_generateSpinCombinations 
TODO optimize this function

CoupledIsingModels.sigmoid — Method

The sigmoid function
note that 
σ(-Inf) ~ 0
σ(Inf) ~ 1

CoupledIsingModels.spinProbabilities — Method

Calculates the probability distribution for each spin independently
idxs says the spins that we care about